Added Data validation, Filtering, Sorting, Pagination

+++ api/Controllers/CommentController.cs
@@ -28,6 +28,13 @@ namespace api.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+
+            //Perform validation
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var comments = await _commentRepo.GetAllAsync();
             
             var commentDto = comments.Select(c => c.ToCommentDto());
@@ -35,9 +42,15 @@ namespace api.Controllers
             return Ok(commentDto); 
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")] //api/Comment/1 :int for validation
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            //Perform validation
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var comment = await _commentRepo.GetByIdAsync(id);
 
             if (comment == null)
@@ -48,10 +61,16 @@ namespace api.Controllers
             return Ok(comment.ToCommentDto());
         }
 
-        [HttpPost("{stockId}")]
+        [HttpPost("{stockId:int}")]
         
         public async Task<IActionResult> Create([FromRoute] int stockId, CreateCommentDto commentDto)
         {
+            //Perform validation
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
            if(!await _stockRepo.StockExists(stockId))
            {
                return BadRequest("Stock does not exist");
@@ -64,9 +83,15 @@ namespace api.Controllers
         }
 
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateDto)
         {
+            //Perform validation
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var comment = await _commentRepo.UpdateAsync(id, updateDto.ToCommentFromUpdate());
 
             if (comment == null)
@@ -79,9 +104,15 @@ namespace api.Controllers
         }
 
 
-        [HttpDelete("{id}")]
+        [HttpDelete("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            //Perform validation
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var commentModel = await _commentRepo.DeleteAsync(id);
 
             if (commentModel == null)
