+++ api/Controllers/StockController.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using api.Data;
 using api.Dtos.Stock;
+using api.Helpers;
 using api.Interfaces;
 using api.Mappers;
 using Microsoft.AspNetCore.Mvc;
@@ -24,15 +25,28 @@ namespace api.Controllers
         }
 
         [HttpGet]
-        public async Task<IActionResult> GetAll(){
-            var stocks = await _stockRepo.GetAllAsync(); //Gets all stocks from the database
+        public async Task<IActionResult> GetAll([FromQuery] QueryObject query){
+
+            //Perform validation
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+            
+            var stocks = await _stockRepo.GetAllAsync(query); //Gets all stocks from the database
             var StockDto = stocks.Select(s => s.ToStockDto()); //Selects all stocks from the database and maps them to the StockDto
             return Ok(StockDto); //Returns a 200 status code with the stocks
         }
 
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id){
+            //Perform validation
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var stock = await _stockRepo.GetAsync(id); //Gets the stock from the database
 
             if(stock == null){
@@ -45,6 +59,12 @@ namespace api.Controllers
 
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto){
+            //Perform validation
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var stockModel = stockDto.ToStockFromCreateDTO(); //Maps the CreateStockRequestDto to the Stock model
            
             await _stockRepo.CreateAsync(stockModel); //Creates the stock in the database
@@ -56,8 +76,14 @@ namespace api.Controllers
 
 
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto stockRequestDto){
+
+            //Perform validation
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
 
             var stockModel = await _stockRepo.UpdateAsync(id, stockRequestDto); //Updates the stock in the database
 
@@ -69,8 +95,14 @@ namespace api.Controllers
         }
 
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")] //
         public async Task<IActionResult> Delete([FromRoute] int id){
+
+            //Perform validation
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             
             var stockModel = await _stockRepo.DeleteAsync(id); //Deletes the stock from the database
